requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pillow==10.3.0
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
